+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 v0.4.0 (*unreleased*)
 ---------------------
 - officially support python 3.10 (:pull:`115`)
+- colorize removed trailing whitespace (:pull:`120`)
 
 v0.3.4 (17 July 2021)
 ---------------------